Novos ajustes do mecanismo de decrementacao das variaveis empilhados na arvore. Correcoes de comportamento da piece var que recebia String em vez de Object no momento de fazer setpiece. correcao de bug na funcao justify.

+++ src/main/java/br/com/innovatium/mumps2java/datastructure/Node.java
@@ -205,7 +205,7 @@ public class Node implements Comparable<Node> {
 		boolean boothNumeric = this.isNumeric && o.isNumeric;
 		if (boothNumeric) {
 			try {
-				return ((Double) susbscript).compareTo((Double) o.susbscript);
+				return ((Integer) susbscript).compareTo((Integer) o.susbscript);
 			} catch (ClassCastException e) {
 				throw new IllegalArgumentException(
 						"There is some inconsistence when was setted the nodes "
@@ -260,9 +260,9 @@ public class Node implements Comparable<Node> {
 		this.stackLevel = stackLevel;
 	}
 
-	private Double isNumber(String string) {
+	private Integer isNumber(String string) {
 		try {
-			return Double.parseDouble(string);
+			return Integer.parseInt(string);
 		} catch (Exception e) {
 			return null;
 		}
